Get a list of active FMI observation stations.

A table of active observation stations is downloaded from the website of Finnish Meteorological Institute, if package rvest or package XML is installed. If neither is, or if the download fails for any other reason, a local copy provided as a csv file within the fmi package is used.

fmi_stations()

Arguments

url
character URL where to look for the weather stations table

quiet
logical whether to suppress printing of diagnostic messages

Value

a data.frame of active observation stations

Details

fmi_weather_stations() is a deprecated alias for fmi_stations(groups="Weather stations").

See also

Author

Joona Lehtomaki joona.lehtomaki@gmail.com, Ilari Scheinin